use lazy table provider only if database fetcher is provided
