+++ src/session.rs
@@ -754,14 +754,13 @@ fn register_catalogs_from_schemas(
             for (table, (schema_ref, batches)) in tables {
                 log::debug!("-- table {:?}", &table);
 
-                let base =
-                    ObservableMemTable::new(table.clone(), schema_ref, log.clone(), batches);
-                // Lazily wrap pg_database if a fetcher is registered.
-                let provider: Arc<dyn datafusion::datasource::TableProvider> = if table == "pg_database" {
-                    crate::lazy_pg_catalog_helpers::wrap_pg_database_provider_if_lazy(Arc::new(base))
-                } else {
-                    Arc::new(base)
-                };
+                let base = ObservableMemTable::new(
+                    table.clone(),
+                    schema_ref,
+                    log.clone(),
+                    batches,
+                );
+                let provider: Arc<dyn datafusion::datasource::TableProvider> = Arc::new(base);
                 schema_provider.register_table(table, provider)?;
             }
         }
